Extract install logic into pegasus_cli/install.py
Move AST-based settings/urls modification functions out of startapp.py
into their own module for better separation of concerns.
